tooltipMui and sliderMui shifted to separate files
