+++ src/lab/Graphs/LineAreaGraph/styles.ts
@@ -1,10 +1,4 @@
-import {
-  makeStyles,
-  Slider,
-  Theme,
-  Tooltip,
-  withStyles,
-} from "@material-ui/core";
+import { makeStyles, Theme } from "@material-ui/core";
 
 interface StyleProps {
   width: number;
@@ -156,57 +150,5 @@ const usePlotLineAreaGraphStyles = makeStyles((theme: Theme) => ({
     strokeOpacity: 0.2,
   },
 }));
-const TooltipMui = withStyles((theme: Theme) => ({
-  tooltip: {
-    backgroundColor: theme.palette.background.paper,
-    color: theme.palette.text.primary,
-    boxShadow: `0rem 0.128rem 0.384rem ${theme.shadows[1]}`,
-    fontSize: "0.9rem",
-    fontWeight: 400,
-  },
-  tooltipPlacementBottom: {
-    margin: theme.spacing(0.5),
-    height: "1.5rem",
-  },
-}))(Tooltip);
 
-const SliderMui = withStyles((theme: Theme) => ({
-  root: {
-    color: theme.palette.highlight,
-    height: "0.375rem",
-    "& .MuiSlider-mark": {
-      display: "none",
-    },
-  },
-  thumb: {
-    height: "0.9rem",
-    width: "0.9rem",
-    backgroundColor: theme.palette.highlight,
-    border: `4px double ${theme.palette.background.paper}`,
-    "&:focus, &:hover, &$active": {
-      boxShadow: "inherit",
-    },
-  },
-  active: {},
-  valueLabel: {
-    left: "calc(-50% + 4px)",
-  },
-  track: {
-    height: "0.375rem",
-    borderRadius: "0.3125rem",
-  },
-  rail: {
-    height: "0.375rem",
-    borderRadius: "0.3125rem",
-  },
-  markLabel: {
-    '&[data-index="0"]': {
-      transform: `translate(0%,0)`,
-    },
-    '&[data-index="1"]': {
-      transform: `translate(-100%,0)`,
-    },
-  },
-}))(Slider);
-
-export { useStyles, usePlotLineAreaGraphStyles, TooltipMui, SliderMui };
+export { useStyles, usePlotLineAreaGraphStyles };
